+++ v2/Changelog.md
@@ -1,6 +1,15 @@
 <!-- markdownlint-disable MD024 -->
 # dar-backup Changelog
 
+## v2-beta-0.6.14 - 2025- -
+
+Github link: [v2-beta-0.6.14](https://github.com/per2jensen/dar-backup/tree/v2-beta-0.6.14/v2)
+
+### Added
+
+- `dar` catalog in xml now parsed correctly, test case added 
+
+
 ## v2-beta-0.6.13.1 - 2025-02-25
 
 Github link: [v2-beta-0.6.13.1](https://github.com/per2jensen/dar-backup/tree/v2-beta-0.6.13.1/v2)